new functions part 2

+++ libft/ft_strlen.c
@@ -1,18 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   strlen.c                                           :+:      :+:    :+:   */
+/*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ide-frei <ide-frei@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/04 19:17:51 by ide-frei          #+#    #+#             */
-/*   Updated: 2022/05/04 19:31:34 by ide-frei         ###   ########.fr       */
+/*   Updated: 2022/05/11 21:07:53 by ide-frei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(char *str)
+#include "libft.h"
+
+size_t	ft_strlen(char *str)
 {
-	int	x;
+	size_t	x;
 
 	x = 0;
 	while (*str)
@@ -27,6 +29,6 @@ int	ft_strlen(char *str)
 int	main(void)
 {
 	char str[] = "hello world!\n";
-	printf("%i\n", ft_strlen(str));
+	printf("%zu\n", ft_strlen(str));
 	return(0);
 }*/
